bug: fixed a weird case of not finding existing controller
Honestly not sure what to make of this one. Elsewhere I was using NotifyOnNewObject to trigger initialization of a module when a new player controller appears then in that initalization, UEHelpers is loaded and GetWorldContextObject is called which gets all player controllers. But for some reason, the way that it was checking if a controller is valid was not working and was printing to the console saying that no player controller was found!